chore: moved historial indstr stats and indst stats getter hook
