+++ src/pages/liquidity-health/index.tsx
@@ -2,8 +2,8 @@ import PageContainer from 'components/PageContainer'
 import LHD from 'views/LHD'
 
 // Hooks
-import { getIndustryStats } from 'hooks/queries/useGetIndustryStats'
-import { getHistoricalIndustryStats } from 'hooks/queries/useGetHistoricalIndustryStats'
+import { getIndustryStats } from 'state/lhd/hooks/useGetIndustryStats'
+import { getHistoricalIndustryStats } from 'state/lhd/hooks/useGetHistoricalIndustryStats'
 import { getLHDProfiles } from 'hooks/queries/useGetLHDProfiles'
 import { getFilterDiff } from 'views/LHD/components/SearchBar/helpers'
