+++ src/state/lhd/types.ts
@@ -1,0 +1,17 @@
+export interface IndustryStats {
+  averageConcentrationScore: number
+  averageHealthScore: number
+  averageOwnershipScore: number
+  averageTotalScore: number
+  chainsSupported: number
+  coefficients: {
+    concentration: number
+    health: number
+    ownership: number
+  }
+  createdAt: string
+  evmCoverage: string
+  formulaVersion: string
+  tokensTracked: number
+  tokensVerified: number
+}
