+++ src/utils/types/lhd.ts
@@ -1,21 +1,3 @@
-export interface IndustryStats {
-  averageConcentrationScore: number
-  averageHealthScore: number
-  averageOwnershipScore: number
-  averageTotalScore: number
-  chainsSupported: number
-  coefficients: {
-    concentration: number
-    health: number
-    ownership: number
-  }
-  createdAt: string
-  evmCoverage: string
-  formulaVersion: string
-  tokensTracked: number
-  tokensVerified: number
-}
-
 export interface ExternalDataOption {
   amount: number
   source: 'cmc' | 'cg' | 'self-reported-cmc'
