+++ src/views/LHD/components/TitleCards/index.tsx
@@ -7,8 +7,8 @@ import useModal from '../../../../hooks/useModal'
 import FilterModal from '../SearchBar/FilterModal'
 
 // Hooks
-import useGetIndustryStats from 'hooks/queries/useGetIndustryStats'
-import useGetHistoricalIndustryStats from 'hooks/queries/useGetHistoricalIndustryStats'
+import useGetIndustryStats from 'state/lhd/hooks/useGetIndustryStats'
+import useGetHistoricalIndustryStats from 'state/lhd/hooks/useGetHistoricalIndustryStats'
 
 // Types
 import { Filters } from 'utils/types/lhd'
